Últimos cambios al código. Hay que intentar reducir el uso de recursos de acceso a la matriz de pesos del SVM como sea.

+++ src/include/dictionary.hpp
@@ -160,7 +160,7 @@ DictionaryEntry<delta_t, confidence_t> Dictionary<index_t, delta_t, confidence_t
 		res.delta = delta;
 		res.valid = true;
 		res.confidence = DICTIONARY_LFU_INITIAL_CONFIDENCE;
-		dictionaryEntries[resultIndex] = res;
+		dictionaryEntries[(int)resultIndex] = res;
 		// this->updateConfidence(dictionaryEntries, resultIndex);
 	}
 	this->updateConfidence(dictionaryEntries, resultIndex);
